Add a wrapper to properly test CTRL-C (SIGINT) on Windows

+++ bin/concurrently.spec.ts
@@ -38,8 +38,12 @@ afterAll(() => {
  * Creates a child process running 'concurrently' with the given args.
  * Returns observables for its combined stdout + stderr output, close events, pid, and stdin stream.
  */
-const run = (args: string) => {
-    const child = spawn('node', [path.join(tmpDir, 'concurrently.js'), ...stringArgv(args)], {
+const run = (args: string, ctrlcWrapper?: boolean) => {
+    const childArgs = ['node', path.join(tmpDir, 'concurrently.js'), ...stringArgv(args)];
+    if (ctrlcWrapper) {
+        childArgs.unshift('fixtures/ctrlc-wrapper/start.exe');
+    }
+    const child = spawn(childArgs[0], childArgs.slice(1), {
         cwd: __dirname,
         env: {
             ...process.env,
@@ -160,23 +164,24 @@ describe('exiting conditions', () => {
     });
 
     it('is of success when a SIGINT is sent', async () => {
-        const child = run('"node fixtures/read-echo.js"');
+        // Windows doesn't support sending signals like on POSIX platforms.
+        // However, in a console, processes can be interrupted with CTRL+C (like a SIGINT).
+        // This is what we simulate here with the help of a wrapper.
+        const child = run('"node fixtures/read-echo.js"', isWindows ? true : false);
         // Wait for command to have started before sending SIGINT
         child.log.subscribe((line) => {
             if (/READING/.test(line)) {
-                process.kill(child.pid, 'SIGINT');
+                if (isWindows) {
+                    // Instruct the wrapper to send CTRL+C to its child
+                    child.stdin.write('^C\n');
+                } else {
+                    process.kill(child.pid, 'SIGINT');
+                }
             }
         });
         const exit = await child.exit;
 
-        // TODO
-        // Windows doesn't support sending signals like on POSIX platforms.
-        // In a console, processes can be interrupted with CTRL+C (SIGINT).
-        // However, there is no easy way to simulate this event.
-        // Calling 'process.kill' on a process in Windows means it
-        // is getting killed forcefully and abruptly (similar to SIGKILL),
-        // which then results in the exit code of '1'.
-        expect(exit.code).toBe(isWindows ? 1 : 0);
+        expect(exit.code).toBe(0);
     });
 });
